Store vector embeddings & details of commit into pg
Once embedded commit message is stored in pg, it can be
queried based on l2 distance algorithm for similarity search.

fetched documents can then be provided to LLM as context
for the user query.

TODO: All of the indexing and querying code is in one ns.
Eventually, this code can be refactored and moved to separate
namespaces.